+++ hardware/NestOAuthAPI.cpp
@@ -616,7 +616,7 @@ void CNestOAuthAPI::SetSetpoint(const int idx, const float temp)
 	}
 
 	int iThermostat = (idx - 1) / 3;
-	if (iThermostat > m_thermostats.size())
+	if (iThermostat > (int)m_thermostats.size())
 		return;
 
 	if (m_thermostats[iThermostat].Serial.empty())
@@ -658,7 +658,7 @@ bool CNestOAuthAPI::SetManualEcoMode(const unsigned char node_id, const bool bIs
 	int iThermostat = (node_id - 4) / 3;
 
 	// Check if we even got that many thermostats.
-	if (iThermostat > m_thermostats.size())
+	if (iThermostat > (int)m_thermostats.size())
 		return false;
 
 	// Grab a reference to that thermostat.
@@ -736,7 +736,7 @@ bool CNestOAuthAPI::PushToNestApi(const std::string & /*sMethod*/, const std::st
 bool CNestOAuthAPI::SetAway(const unsigned char Idx, const bool bIsAway)
 {
 	int iStructure = (Idx - 3) / 3;
-	if (iStructure > m_structures.size())
+	if (iStructure > (int)m_structures.size())
 		return false;
 
 	if (m_structures[iStructure].StructureId.empty())
